Exercícios de testes de erro

+++ src/business/UserBusiness.ts
@@ -1,5 +1,5 @@
 import { UserDatabase } from "../database/UserDatabase"
-import { DeleteUserInputDTO, GetAllOutputDTO, GetUserInputDTO, LoginInputDTO, LoginOutputDTO, SignupInputDTO, SignupOutputDTO } from "../dtos/userDTO";
+import { DeleteUserInputDTO, GetAllOutputDTO,  GetUserInputDTO, LoginInputDTO, LoginOutputDTO, SignupInputDTO, SignupOutputDTO } from "../dtos/userDTO";
 import { BadRequestError } from "../errors/BadRequestError";
 import { NotFoundError } from "../errors/NotFoundError";
 import { User } from "../models/User";
@@ -69,7 +69,6 @@ export class UserBusiness {
 
         return output
     }
-
     public login = async (input: LoginInputDTO): Promise<LoginOutputDTO> => {
         const { email, password } = input
 
@@ -139,24 +138,34 @@ export class UserBusiness {
         return output
     }
     public deleteUser = async (input: DeleteUserInputDTO): Promise<void> => {
-        const { id } = input
+        const { id, token } = input
+        if (typeof token !== "string") {
+            throw new BadRequestError("requer token")
+        }
+
+        const payload = this.tokenManager.getPayload(token)
+
+        if (!payload) {
+            throw new BadRequestError("token inválido")
+        }
+
+        if (payload.role !== USER_ROLES.ADMIN) {
+            throw new BadRequestError("somente admins podem deletar contas")
+        }
+
         const userDeleteDB = await this.userDatabase.findUserById(id)
         if (!userDeleteDB) throw new NotFoundError("Id não encontrado")
 
         await this.userDatabase.deletedUserById(id)
     }
 
-    public getUserById = async (input: GetUserInputDTO): Promise<UserDB | undefined> => {
+    public getUserById = async (input: GetUserInputDTO):Promise<GetUserInputDTO>=> {
         const { id } = input
         const user = await this.userDatabase.findUserById(id)
-        if (!user) throw new NotFoundError("Id não encontrado")
-
-        return user
+        if (!user) {
+            throw new NotFoundError("Id não encontrado")
+         }
+         
+        return user 
     }
 }
-
-/*Crie os seguintes endpoints:
-- DELETE /users/:id- que deleta um user específico
-- GET /users/:id- que busca por um user específico e retorna todas as suas informações no modelo de regra de negócio (UserModel)
-
-Ambos devem ser protegidos (requer token).*/
